requirements.txt
pymupdf
sentence-transformers
faiss-cpu
gradio
duckduckgo-search
python-dotenv
google-genai
requests